Added ALL APIs, few bugs

+++ src/main/java/Controllers/Checked.java
@@ -16,16 +16,17 @@ public class Checked {
     @Consumes(MediaType.MULTIPART_FORM_DATA)
     @Produces(MediaType.APPLICATION_JSON)
     public String insertChecked(
-            @FormDataParam("CheckedID") Integer CheckedID){
+            @FormDataParam("CheckedID") Integer CheckedID, @FormDataParam("CheckedCheck") Boolean CheckedCheck){
         try {
-            if (CheckedID == null) {
+            if (CheckedID == null || CheckedCheck == null) {
                 throw new Exception("One or more form data parameters are missing in the HTTP request.");
             }
             System.out.println("thing/new CheckedID=" + CheckedID);
 
-            PreparedStatement ps = Main.db.prepareStatement("INSERT INTO Requirements (CheckedID) VALUES (?)");
+            PreparedStatement ps = Main.db.prepareStatement("INSERT INTO Requirements (CheckedID, CheckedCheck) VALUES (?, ?)");
             ps.setInt(1, CheckedID);
-            ps.executeUpdate();
+            ps.setBoolean(2, CheckedCheck);
+            ps.execute();
             return "{\"status\": \"OK\"}";
 
         } catch (Exception exception) {
@@ -42,12 +43,13 @@ public class Checked {
         System.out.println("Checked/list");
         JSONArray list = new JSONArray();
         try {
-            PreparedStatement ps = Main.db.prepareStatement("SELECT CheckedID FROM Checked");
+            PreparedStatement ps = Main.db.prepareStatement("SELECT CheckedID, CheckedCheck FROM Checked");
 
             ResultSet results = ps.executeQuery();
             while (results.next()) {
                 JSONObject item = new JSONObject();
                 item.put("CheckedID", results.getInt(1));
+                item.put("CheckedCheck", results.getBoolean(2));
                 list.add(item);
             }
             return list.toString();
@@ -59,30 +61,75 @@ public class Checked {
     }
 
 
-    public static void updateChecked (int CheckedID){
+    @POST
+    @Path("update")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    public String updateChecked(
+            @FormDataParam("CheckedID") Integer CheckedID, @FormDataParam("CheckedCheck") Boolean CheckedCheck){
         try {
+            if (CheckedID == null || CheckedCheck == null) {
+                throw new Exception("One or more form data parameters are missing in the HTTP request.");
+            }
+            System.out.println("thing/update CheckedID" + CheckedID);
+            PreparedStatement ps = Main.db.prepareStatement("UPDATE Checked SET CheckedCheck = ? WHERE CheckedID = ?");
+            ps.setBoolean(1, CheckedCheck);
+            ps.setInt(2, CheckedID);
+            ps.execute();
+            return "{\"status\": \"OK\"}";
 
-            PreparedStatement ps = Main.db.prepareStatement("UPDATE Checked SET CheckedID = ?");
-            ps.setInt(1, CheckedID);
+        } catch (Exception exception) {
 
-        } catch (Exception e) {
-
-            System.out.println(e.getMessage());
+            System.out.println("Database error: " + exception.getMessage());
+            return "{\"error\": \"Unable to update item, please see server console for more info.\"}";
 
         }
 
     }
 
-    public static void deleteChecked(int CheckedID) {
+    @POST
+    @Path("delete")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    public String deleteChecked(@FormDataParam("CheckedID") Integer CheckedID) {
         try {
+            if (CheckedID == null) {
+                throw new Exception("One or more form data parameters are missing in the HTTP request.");
+            }
+            System.out.println("thing/delete CheckedID=" + CheckedID);
 
             PreparedStatement ps = Main.db.prepareStatement("DELETE FROM Checked WHERE CheckedID = ?");
             ps.setInt(1, CheckedID);
-            ps.executeUpdate();
+            ps.execute();
+            return "{\"status\": \"OK\"}";
 
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
+        } catch (Exception exception) {
+            System.out.println("Database error: " + exception.getMessage());
+            return "{\"error\": \"Unable to delete item, please see server console for more info.\"}";
+        }
+    }
 
+    @GET
+    @Path("get/{CheckedID}")
+    @Produces(MediaType.APPLICATION_JSON)
+    public String getChecked(@PathParam("CheckedID") Integer CheckedID) {
+        if (CheckedID == null) {
+            throw new Exception("Checked's 'CheckedID' is missing in the HTTP request's URL.");
+        }
+        System.out.println("thing/get/" + CheckedID);
+        JSONObject item = new JSONObject();
+        try {
+            PreparedStatement ps = Main.db.prepareStatement("SELECT CheckedCheck FROM Checked WHERE CheckedID = ?");
+            ps.setInt(1, CheckedID);
+            ResultSet results = ps.executeQuery();
+            if (results.next()) {
+                item.put("CheckedID", CheckedID);
+                item.put("CheckedCheck", results.getBoolean(1));
+            }
+            return item.toString();
+        } catch (Exception exception) {
+            System.out.println("Database error: " + exception.getMessage());
+            return "{\"error\": \"Unable to get item, please see server console for more info.\"}";
         }
     }
 }
